requirements.txt
Django==4.2
djangorestframework==3.14.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0